# static/docs/swagger.yaml
definitions:
  models.ApiKey:
    properties:
      apiKey:
        type: string
    type: object
...
  models.Email:
    properties:
      email:
        type: string
    type: object
  models.HasData:
    properties:
      hasData:
        type: boolean
    type: object
...
info:
  contact:
    email: me@dunkirk.sh
    name: Kieran Klukas
    url: https://github.com/kcoderhtml
  description: |-
    REST API to interact with [Hackatime](https://w.dunkirk.sh)

    ## Authentication
    Set header `Authorization` to your API Key encoded as Base64 and prefixed with `Basic`
    **Example:** `Basic ODY2NDhkNzQtMTljNS00NTJiLWJhMDEtZmIzZWM3MGQ0YzJmCg==`

    If you are using the admin token then set header `Authorization` to the admin token prefixed with `Bearer` and not encoded with anything
    **Example:** `Bearer blahaji_rulz_da_world` One point to be aware of is that when using this method you always need to pass the `user` query parameter with the id of the user you wish to query for.
  license:
    name: MIT
    url: https://github.com/kcoderhtml/hackatime/blob/master/LICENSE
  title: Hackatime API
  version: "1.0"
paths:
...
  /special/apikey:
    get:
      operationId: get-apikey
      parameters:
      - description: The user to filter by if using Bearer authentication and the
          admin token
        in: query
        name: user
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ApiKey'
      security:
      - ApiKeyAuth: []
      summary: Retrieve a users API key
      tags:
      - apikey
  /special/email:
    get:
      operationId: get-email
      parameters:
      - description: The user to filter by if using Bearer authentication and the
          admin token
        in: query
        name: user
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Email'
      security:
      - ApiKeyAuth: []
      summary: Retrieve a users email
      tags:
      - email
  /special/hasData:
    get:
      operationId: has-data
      parameters:
      - description: The user to filter by if using Bearer authentication and the
          admin token
        in: query
        name: user
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.HasData'
      security:
      - ApiKeyAuth: []
      summary: Whether the user has data any heartbeats received yet
      tags:
      - hasData
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"